Add a fast-forward test
I modified this test slightly, but it's generally what Clark wrote
in https://review.opendev.org/784142

That test fails against master, but this does not:

I don't think the Zuul ref check at the end is necessary, and I
added a checkoutBranch call, since what we really want to do
is make sure that we get the correct thing checked out.

With those changes, along with a corrected form of the repo state
dict, this passes on master and can be used as a regression test
for the global repo state work.

Note, however, that this only passes due to the mismatch of the
repo state dict format.  That will be corrected in a followup
change.

Co-Authored-By: Clark Boylan <clark.boylan@gmail.com>

Change-Id: I7f01910f189984f6bd80c09a2af2ec3536d102fe