Correct repo_state format in isUpdateNeeded
The isUpdateNeeded call was not operating on the actual dictionary
format that is passed in.  The tests did not catch this because they
pass in the format that is expected.  Update both the tests and the
calling code in the merger to fix.

This breaks the just-added fast-forward test, which shows us that
the current behavior really is broken.

Change-Id: I34b7dbe1d4f7032d217bca30ca9a8d3c986c1915