Restore repo state in checkoutBranch
If we pass a repo state to checkoutBranch, we should check out the
repo/branch as specified by the repo state.

FYI, there are two theoretical ways we could call checkoutBranch:

1) Without a repo state (in which case we would expect to checkout
the current value of the branch in the repo; presumable this would
be done after an update() call also without a repo state, which means
the current branch would match upstream).

2) With a repo state, in which case we would expect to check out
the branch as specified by the repo state.

Currently Zuul only uses option 2.

Change-Id: Icad68a337b3ff5fc80af32ee0a4845cc83daa14b