Update patch set 1
Patch Set 1:

I believe this only works because we're giving it realistic data; Clark's original test gave it incorrectly formatted data (in the same incorrect format as the isUpdateNeeded tests), and that's part of why it fails against master.  By "correcting" the data in this patch, isUpdateNeeded will *always* return True, which will cause things to work.

Clearly that's not what we want, so the subsequent patches work through the process of getting us to a place where isUpdateNeeded is called with the correct data and works as intended.

Patch-set: 1