Pseudo-shard unique project names in keystore
There is a 4mb (by default) limit to the size of a packet which
can be sent to a zk client, which means with a typical repo name
length of around 25 chars we could store somewhere between
100,000 or 200,000 projects per connection.

An easy way to potentially increase that number is to shard by
path component, since most connections that would have that many
repos probably also have paths which would provide useful sharding.

This makes the keystore project path like:

  connection/prefix/project

Where prefix is the first path component. The project is still the
full project name with url quoting, so there is no issue with
collisions.

Some examples:

  project                -> project/project
  project/subproject     -> project/project%2Fsubproject
  project/sub/subproject -> project/project%2Fsub%2Fsubproject

Change-Id: I59a6e04a0520b65c4e5a5a87a38e2db0216ffc30