Support key versions and unique names in ZK keystorage
The new ZK keystorage did not support versioned keys, which is
something we did support in the file storage.  To make the new
ZK keystorage forward-compatible with multiple keys, turn the
stored JSON structure into a list in a dict.

The outer dict lets us store information about the entire
blob (for example, the schema of the data it contains) and allows
us to add any future attributes.

The 'keys' attribute of the outer dict is a list of key structures,
each of these corresponding to the data stored previous to this
patch.  We also add some other useful attributes, such as the key
version (always 0 for now) and the created timestamp (which may be
the imported time if we import from backup).

In the future, key rotation may be accomplished by appending to
the list of keys and then removing old keys.  I considered making
each key a distinct znode, but chose to store all keys within a
single znode as it will simplify key loading operations during
rollover (where we will need to load all keys).  It will be a single
read instead of a get_children+N*get.

We omit storing the public key from the secrets store in order
to match the behavior of the current filestore and the ssh keystore
(the public key is redundant since it is derived from the private
key).

Also, modify the project name to be a single unique string.
The concern is a system where there is both a "project" and a
"project/subproject" repo.

While this is technically not a problem with ZK because a znode
can hold both data and children, it would make things less
ambiguous if we had a hierarchy of known depth in ZK and data
nodes were not children of other projects data nodes.  For example,
imagine deleting "project" from the system without deleting
"project/subproject".  This is much simpler if subproject is
not a child znode of project.

Or consider the case of a project literally named "project/secrets".
Again, technically not a problem in ZK (project/secrets holds the
secrets for project, and project/secrets/secrets holds the secrets
for project/secrets).  But it would make anyone debugging it tear
their hair out (if any remains).

To resolve this, we use urllib's quote_plus function.  This is already
in use for the semaphores in ZK. This makes the project name
predictable, flat, unique, and mostly human-readable.

E.g., 'project%2Fsubproject' for 'project/subproject'.

No migration is provided for the ZK changes since this patch series
has not merged yet.

Change-Id: I13dffa4d9592a773172e07c9b507fac8a9a2335d