Don't check for existing refs in isUpdateNeeded
The isUpdateNeeded function is called to check if a repo state could
be restored on a repo without prior fetching. This only requires
checking for the revs that are part of the repo state since any
reference would be deleted and re-created during repo state restore as
well. Thus remove the check for the existing refs.

Change-Id: I3c61e52dcbb8d24cebda5a79b5c2d6f08ef25140