Cache secret/SSH keys from Zookeeper
Loading secret/SSH keys from Zookeeper can be expensive in case a tenant
contains a lot of projects. Since the keys were also loaded from
Zookeeper during reconfigurations, this can significantly increase the
reconfig time (around +30s for 500 projects).

In order to cache the key we need to instanciate a single keystore
instance in the scheduler instead of creating a new one for each config
loader.

Change-Id: I377777907ce325d5aaac7b18db1d3eda499d6b4f