Add ExecutorApi
This change implements a ZooKeeper ExecutorApi that stores build
requests in ZooKeeper.

The build requests are stored in the path:
    /zuul/build-requests/<zone>/<uuid>

A later change will switch the executor to use it.  The intention is:

The executor client puts new build requests in ZooKeeper and can request
cancel and resume actions via child nodes.

The executor server listens for new requests via a children watch on
the build requests path.

Cancel and resume requests can be scheduled via child nodes on a build
request.

This change still relies on local Build objects being present on the
scheduler in addition to the BuildRequests being stored in ZooKeeper.

The BuildRequest is removed together with the local Build object when
the scheduler has finished the result event processing.

Change-Id: I5de26afdf6774944b35472e2054b93d12fe21793