Change zone handling in ExecutorApi
This changes how zones are handled in the ExecutorApi.

This creates an unambiguous unzoned root, as well as a root for
each zone.  Unzoned jobs are represented with None as a zone.
This allows us to remove a complicated method of setting dynamic
watches.

This also adds a test which simulates both sides of the API
(scheduler and executor) so that we can exercise the full lifecycle
of a request easily and to show concisely how the API should be
used.

Change-Id: I624f23b47daa5273b69289a6c7a2187c7ba9ddde