Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> PS5 is a non-functional rebase on current master which includes the global repo state changes.  It functionally conflicts, so I have resolved the textual conflicts with comments for now and will work on a new patch to resolve the functional conflicts.

I was also working on this rebase on Friday, but didn't manage to finish it. To be honest I didn't expect you to also do this rebase. Maybe next time I should set a Workflow -1 on my change when I'm still working on it although it has merge conflicts.

I went a slightly different approach and "fixed" the missing parts in the cleanup functions (deleting the build, returning the nodeset, releasing the semaphore) by directly processing a faked BuildCompletedEvent. I thought that this is the part which is now different with the result event queues being put into ZooKeeper. To not conflict with your rebase, I pushed this change here https://review.opendev.org/c/zuul/zuul/+/787915. However the `tests.unit.test_github_driver.TestGithubDriver.test_timer_event` test is failing with my appraoch, although I'm not sure if that is really a problem or if we just have to adapt this test.

I'm not sure if my version is a proper solution or should we follow on your version with the force flag. I think your version separates the cleanup from the normal workflow a little better.

I don't mind which change we use, there is just one difference between those two which we should include in your change, if we are going to use it. This is the part where the actual cleanup of the tenant in ZooKeeper is done. This part should be put at the end of the _reconfigureDeleteTenant() function in the scheduler:

# Delete the tenant event root path for this tenant in ZooKeeper to
# remove all tenant specific event queues
try:
    self.zk_client.client.delete(
        f"{TENANT_EVENT_ROOT}/{tenant.name}", recursive=True
    )
    except NotEmptyError:
        # In case a build result has been submitted during the
        # reconfiguration, this cleanup will fail. We handle this in a
        # periodic cleanup job.
        pass

Patch-set: 6