Update patch set 6
Patch Set 6:

I just saw that the tenant cleanup in _reconfigureDeleteTenant is already there. We should just wrap it into a try/except block to catch the possible NotEmptyError in case a build result was put in the queue during the reconfiguration/delete:

try:
    self.zk_client.client.delete(
        f"{TENANT_EVENT_ROOT}/{tenant.name}", recursive=True
    )
    except NotEmptyError:
        # In case a build result has been submitted during the
        # reconfiguration, this cleanup will fail. We handle this in a
        # periodic cleanup job.
        pass

We should then use this change over mine since it's fully functional.

Patch-set: 6