WIP Add some executor API tests
If the scheduler deletes a locked build request (which we may have
it do if we delete a tenant) we still need the executor to abort
the job.  Simulate that behavior with the ExecutorApi.

Add a test of the hold functionality.

If a request is removed right before the executor locks it, make
sure that the executor knows it did not lock it.

Change-Id: If3364e5c045c1704fb28a30e38d8c6086183089a