Clean up Executor API build request locking and add tests
If the scheduler deletes a locked build request (which we may have
it do if we delete a tenant) we still need the executor to abort
the job.  Simulate that behavior with the ExecutorApi.

Add a test of the hold functionality.

Delete the lock parent node when we delete the build request
(otherwise we will leak zk nodes).

If a request is removed right before the executor locks it, make
sure that the executor knows it did not lock it, and also clean
up the lock.

Change-Id: If3364e5c045c1704fb28a30e38d8c6086183089a