getHoldRequest: pad zeros if not full length
As noted inline, we know that the autohold sequential id's should be
zero padded to 10 digits.  In the autohold-list you're usually seeing
a very small id like "0000000184"; when you try to do the logical
thing of "zuul autohold-delete 184" this silently fails.

Make this a bit less of a trap by zero padding id's that are less than
10 digits.

Change-Id: I4272919fba3571c27d19f28f7dba64edb99c3adf