Optimize layout re-calculation after re-enqueue
An item's layout is mainly required for freezing the job graph and in
case a following item needs it as a fallback layout.

Since we no longer re-freeze the job graph during re-enqueue as of
I2e4bd2fb9222b49cb10661d28d4c52a3c994ba62, we can save some time by not
re-calculating the layout for items that already have a job graph.
Instead we only re-generate the 'nearest' fallback layout on demand in
case a following item needs it.

This also prepares us for the scale-out scheduler where we have to
re-calculate a layout on a scheduler if it is not available locally.

Change-Id: Id7cef4f1fa222b1491418ea2449687964fcfb361
Co-Authored-By: Simon Westphahl <simon.westphahl@bmw.de>