Calculate build start and end time on executor server
This sets the build's start and end time on the executor server rather
than the scheduler. This is necessary to calculate the in-use duration
of locked nodes later on when the nodes are locked on the executor.

Since the scheduler must still set the start and end time on the local
build object, we use the BuildStarted and BuildCompleted to provide the
timestamps to the scheduler. Once the builds are stored in ZooKeeper,
the executor can directly set the timestamps on the build.

In case of noop builds the timestamps will be set by the executor client
as those builds don't show up on any executor server.

Change-Id: I08ac9618b3c12280a8720657f97d0577803e1a59