Always request all required project configuration
In preparation for caching a project's config in Zookeeper we need a way
to request all necessary config for all tenants the project is part of
with a single merger call.

To do that, we create the list of all TPCs before we start loading any
tenant. We can then combine the list of extra config files/dirs for a
project from all tenants for the cat job.

It's worth noting that this change adds some overhead for projects using
extra config files/dirs without providing any immediate benefits.

Change-Id: If0d134f2583ba5a5a176ba29f6d9f906b46aad05