Update patch set 20
Patch Set 20:

(3 comments)

I have split this up into a series of changes starting at https://review.opendev.org/787832.  It adds the API first, then changes it a bit to address some of my comments here about the zones and the unused "use_cache" argument.  It also adds tests of the API so we can see it used and more easily verify behaviors.  Then it actually switches over to using it and updates the rest of the tests.  I addressed some of my own comments, and I also made some minor changes which I didn't comment.  I'm sorry about that -- it's not the best way of working -- but I didn't see some things until I started working on it.

I did not increase the test run-time because I don't think that we should need to do that; instead, I altered the sleep you added (though I think we should work to remove that).  I reduced the differences between the "test" version of the executor API and the real one.  We should try never to alter the behavior of the classes in tests.  When we do change them, we try to only add additional methods, or wrap the original ones with additional behavior.  Finally, I removed the cleanup thread from the executor since I think it's the wrong place, but I have not yet added it to the scheduler, which I think is the right place.

I think this way of structuring the ExecutorAPI is the best way to go right now.  I tried out several other ideas (privately; I didn't push any code up), and I wasn't as happy with any of them as with this.  I think we do want to keep using a cache on the object.  With some changes, we *could* eliminate the use of the cache on executors (use a priority queue numbering scheme for the build requests), however, the cleanup thread in the executor would still need to iterate over all of the build requests in order to determine which are lost.  It's better to use a cache for that, and in that case, we may as well use it on the executor too.

Once the pipeline state is in ZK, we might be able to remove build requests after they start, which might open up other options, but that's too many steps away.

Once thing I notice is that it looks like we're passing the decrypted secrets via build requests, which violates our "encrypted at rest" rule for ZK, so we'll need to have the executors do secret decryption.

Here's how I think we should proceed:

1) Before anything else, write a new change to have the existing gearman api send *encrypted* secrets to the executors, and have the executors decrypt them by reading the keys from ZK.  It should also not send the ssh keys, and instead the executors should read the ssh keys from ZK and decrypt them.  (Note, with the executor client no longer handling secrets, we may actually be able to remove the redaction that we do in executor/common.py.  But we can leave that as-is for now.)

2) Reconcile this change with my stack.  Either switch to my stack, or rework this change to be equivalent to the first change in my stack and put the rest of my stack on top of it, or whatever seems best.  I do want to make this at least 2 changes: add the API, then using it.

3) Add the cleanup thread to the executor.

Patch-set: 20
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>