Fix resolveBranch working on outdated data
Change I740e958357dc7bf0a6506474c5991da12ab6264e refactored the
repo.update and repo.reset methods such they do what the name says. In
this regard the repo reset is not executed anymore on
repo.update. This is correct, however the executor then takes an
outdated list of branches and refs from the repo.update call. This can
lead to unexpected silent fall backs to the project default branch in
case the branch is missing in the local branches of the cached repo.

To fix this feed resolveBranch with the refs from the repo state
instead which should be treated as the source of truth.

Change-Id: I340dfdac9a730dbf7a562202e925c058873a8e73