Fix race in test for cancel of starting builds
Since the worker info required for canceling a build is now set in the
run handler when the BuildStartedEvent is processed, we need to also
wait for the worker info to be available.

Otherwise the event for the abandoned change might be processed before
the BuildStartedEvent, which will cause the test to fail.

Change-Id: I3128e97dfa5f40667ede274bd8718b1a0a5815a6