Add a logical timestamp to management events
Create a 'zuul_event_ltime' attribute for management events, which is
the Zookeeper creation transaction ID of event's Znode. In case the
event's ltime is already set it won't be change, e.g. for forwarded
events.

This timestamp can e.g. be used later on to determine if a projects
config cache in Zookeeper is up-to-date (happend before relation).

Change-Id: I922fef7ddc833565a97abe798fdecdb22b0d8dcb