Implement API for caching raw config files in ZK
The content of config files will be saved in Zookeeper using the
following path schema:

    /zuul/config/cache/<project-cname>/<branch>/<filename>

The canonocal project name, branch and filename are escaped using 'plus
quote' to avoid issues with '/' as part of the names.

The list of extra config files/dirs the cache is valid for will be
stored in the root node of the project-branch cache.

In order not to store any duplicated data in Zookeeper, the unparsed
files cache is global and shared by all tenants. Because of that it
needs to be read/write locked on a per project basis. Locks will be
stored separate from the cache content in order to avoid any conflicts
e.g. when a branch cache is deleted:

    /zuul/config/lock/<project-cname>

As we need a consistent view of the unparsed config, Zookeeper must be
accessed directly instead of using a TreeCache. Unfortunately, with the
TreeCache there is no reliable way to determine if ALL changes of a
(sub-)tree have been synced to the cache. This is the case as the tree
cache uses data watches internally, which are created as the sync
progresses. Because of that we don't have a guarantee that tree events
are totally ordered.

Change-Id: I215eb26774cd0d53892e6a152bc1011fb72193da