Implement API for caching raw config files in ZK
Unparsed config files will be saved in Zookeeper using the following
path schema:

    /zuul/config/cache/<project-cname>/<branch>/<filename>

The canonocal project name, branch and filename will be 'plus quoted' to
avoid issues with '/' as part of the names.

In order not to store any duplicated data in Zookeeper, the unparsed
files cache is global and needs to be read/write locked on a per project
basis. Locks will be stored separate from the cache content in order to
avoid any conflicts e.g. when a branch cache is deleted:

    /zuul/config/lock/<project-cname>

As we need to a consistent view of the unparsed config we have to
access Zookeeper directly instead of using a TreeCache. Unfortunately,
with the TreeCache there is no reliable way to determine if ALL changes
of a (sub-)tree have been synced to the cache. The reason for this is
that the tree cache uses data watches internally, which are created as
the sync progresses. Because of that we don't have a guarantee that tree
events are totally ordered.

Change-Id: I215eb26774cd0d53892e6a152bc1011fb72193da