Cache unparsed config files in Zookeeper
Cache raw config files in Zookeeper so we can load the config on
multiple schedulers with only one round trip to the mergers.

In case no cache ltime is give, the configloader will get the content of
the config files via a cat job and store the returned config in
Zookeeper.

For tenant reconfigurations the 'zuul_event_ltime' of the corresponding
event will be passed as the cache ltime to 'reloadTenant()'. This has
the advantage that in case a project is part of multiple tenants, only
the first tenant will request the config from the mergers. All following
tenants can get the config from Zookeeper.

In a later step we can save the logical timestamp of a tenant
reconfiguration in the 'global layout state'. This timestamp can then be
used as the cache ltime by other schedulers to update their local
layout.

Change-Id: Ia3d2a124eaf6ee9d82c94a0375ee7ecba61caadf