Support serializing encrypted secret objects
In memory, we store secrets as deserialized YAML objects, which
can be dicts, lists, scalars, or our own custom EncryptedPKCS1_OAEP
subclass of YamlObject.  We need to be able to serialize a
(potentially) encrypted secret in order to transmit it to the
executor via ZK.  The easiest and most sensible serialization
mechanism is to reverse our deserialization and use YAML.

This change supports this by indicating that an EncryptedPKCS1_OAEP
object is safe to serialize using the SafeDumper.  Any time we
safe_dump to yaml a Python data structure with an EncryptedPKCS1_OAEP
in it, it will be serialized using our custom tag.

This also corrects an unrelated parameter name typo in the
safe_dump method.

Change-Id: I1034946e5261003d9e83119ceee99b89d0dc53a1