Allow monitoring length of connection event queues
In order to expose the length of the connection event queues we need to
introduce `getEventQueue()` as a new interface in the connection base
class.

The queue length will be available in the status JSON output for display
in zuul-web as well as a statsd gauge metric:

    zuul.scheduler.eventqueues.connection.<connection-name>

This gives an operator more insight into possible bottlenecks during
event pre-processing.

Change-Id: Ie2e09e0c7cd166dce5b884c3e3d12ead42b05397