Update patch set 10
Patch Set 10:

(5 comments)

> I think I never used autoholds yet, so I don't know much about them. But if it's something we should show to the user, I'm in favor of this change 😉

On the deployments we manage, users occasionally request a autohold to debug a failing job. It's not always easy for them to follow up after the request, especially figure out which build was held if any. Adding these pages will help them know if their request was validated, if it triggered a hold, and if so, which build was held!

Patch-set: 10