Add button to go to top of build page
As a frequent user of the Zuul web UI, I have often wished there were
a way to jump to the top of the build page after scrolling down a
large log file.

This adds a button in the bottom right corner that jumps to the top of
the page when clicked.

Change-Id: I7782055d64255844b682e282fcd1d74cf88f4ced