Raise exception on duplicate tenant configs
If an admin specifies two tenant blocks in main.yaml with the same
name, raise an exception.

Zuul assumes tenant names are unique, and there is no "overlay"
functionality (like we have with jobs, etc) in the main tenant config
file that would warrant two entries with the same name.

Change-Id: I0e44ef644e17f8d412e82ce4ca51febc7c65a00c