Update patch set 13
Patch Set 13:

I mentioned this in IRC but thought I would mention it here so it doesn't get lost.

I'm worried that people won't go looking for the autohold button on build pages because build pages are where you go to see things that have completed and not to change future state.

One idea is the job page? But even that doesn't seem quite right as you can do a hold for all jobs against a change I think.

Patch-set: 13
CC: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>