Update patch set 13
Patch Set 13:

> Patch Set 13:
> 
> I mentioned this in IRC but thought I would mention it here so it doesn't get lost.
> 
> I'm worried that people won't go looking for the autohold button on build pages because build pages are where you go to see things that have completed and not to change future state.
> 
> One idea is the job page? But even that doesn't seem quite right as you can do a hold for all jobs against a change I think.

my reasoning was that a user would go through the following scenario: 

  1. Build for job J on change A fails 
  2. I don't have enough info in the logs/artifacts for a fix and need to do a post-mortem 
  3. I want to hold the nodes next time a build of J fails on change A

In that spirit it makes sense to add the autohold request button on the failed build's page since that's where the decision to hold is made, and all the required arguments for the request can be easily pre-filled this way.

We could also add a "create autohold request" button to the new autoholds page in https://review.opendev.org/c/zuul/zuul/+/768199 - that'd pop a form where you have full control on the query parameters.

Patch-set: 13