Update patch set 5
Patch Set 5:

Is everyone aware this is how files used to work before we changed it, largely at the behest of tripleo?

That doesn't make this any more or less useful -- obviously the extending file matchers was useful before, and the overwriteable file matchers are useful now for different reasons.  It could certainly make sense to have both.

But it's important to know the history[1] to make sure that we don't get into a situation where people start using this, and then decide that they need to override always-files set in different repos and we make a new way of doing that and then end up with always-always-files.

Before we even get that far, there's the obvious complement at this level: always-irrelevant-files.  We don't even have to go more than one level deep before we have four different file-matching attributes.  Some of us have trouble working out the interaction with two of them, much less four.  This could get rather confusing for users.

It's worth considering whether this is the most straightforward solution, or if we should add some more generic way to indicate whether we want to override or extend a job attribute.  For example, we could make "files" extend by default and add a YAML tag to break the inheritance chain.  Then "files: [foo]" would extend the parent's list with foo (like this change), but "files: !override [foo]" would clear it and start anew with only foo (like the current behavior).  (Or of course we could implement the opposite with !extend.)

That would allow quite a bit of the expressiveness that this change achieves, and could be applied to a whole host of job attributes, which may allow us to express more options without users having to remember a lot of new attributes.

[1] http://lists.openstack.org/pipermail/openstack-dev/2018-May/130075.html

Patch-set: 5
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>