Be explicit about yaml loaders/dumpers
In preparation for expanded use of the encrypted YAML object in the
executor, make the yaml loading and dumping more explicit.

We currently rely on metaclass magic to cause the encrypted yaml
objects to work.  To avoid any future potential import errors and
also to make it easy and clear about which kind of objects should
be supported while loading/dumping yaml, stop using the metaclass
method and instead explicitly create encrypted loader/dumper clasess
and helper methods.

Change-Id: I97a65e3f162ffaa20bb09a09278b00a76e7b1c0c