Report unzoned executor stats even at 0
We report online mergers and executors if we see their functions,
but there is an edge case if there are no functions registered.
In that case, since we don't see the gearman function registered,
we don't report the value.  Therefore, if a scheduler comes online
and zero executors join the cluster, we won't report any statsd
value, and statsd will continue to use the last value of the gauge.

A site could configure statsd to delete the gauge after the flust
interval, but we would have to guarantee that we send every gauge
every 10 seconds, and we can't make that guarantee.

We can improve this for at least the unzoned executors by always
at least emitting a zero value for those metrics.  Unfortunately
we can't do the same for zoned executors since we don't know the
whole set of possible zone names.

I think the work in progress to use prometheus for availability
metrics is a solution for this, in either of two ways:

1) The prometheus system itself can be configured to scrape every
   executor and therefore will detect when any or all are offline.
2) If Zuul provides a centralized endpoint (i.e., a prom endpoint
   where we report stats like zuul_executors_online), it can report
   metrics for all the zones and therefore prometheus alerting
   rules can take that into account.  In other words, Zuul will
   simply not supply a metric for a zone it doesn't know about,
   but a prometheus rule can be written for that zone if the
   site admin knows it should exist.  Prometheus can detect that
   the metrics endpoint is up but not reporting a particular zone,
   so the case of a zone that Zuul doesn't know about can be handled.

Change-Id: Idd8f76935fd9ccaa73800b8ae7ab038f5dc83995