tutorial: Export log, Update container name
docker-compose in focal doesn't handle container name with '-' the same
way as the version included in bionic.
Update the container name to be correctly handled using the newest
version of docker-compose.

This will match recent update for opendev infra, people localy running
tutorial with an old docker-compose will face an issue, as this export
is mainly useful for opendev and is now flagged with
"ignore_errors: yes", it won't have a significant impact for those users

Change-Id: I7b4221c1446b071f2b3a86af7d67ad55f92a68a1