Fix race with watches in ExecutorAPI
The ExecutorAPI sets watches on build requests when they are
created.  When a watch is created using the kazoo DataWatch recipe,
it performs an initial callback without an event, which can be used
to prime a cache.  The executor api did not take advantage of it
because it primes its cache before setting the watch.  However,
Data in ZK can change between the "get" call that collects the data
and the "get" call that sets the watch.  This can even happen in our
unit tests; the test_lost_build_requests test is capable of triggering
this race.

An obvious solution is to prime our cache on the initial callback.
However, kazoo does not supply the path in that case (we get it from
the event object, which is missing on the inital call).  To resolve
this, a function factory method is added to return a closure for the
callback with the path as part of the closure.  Now our callbacks always
get a path argument.

Everything above also applies to the ChildrenWatch that we set to be
alerted of new build requests.  This also corrects a similar/related
issue in that if an executor came online with existing build requests in
ZK, it would not notice them until a new build request triggered the
watch.  Applying this same fix causes a "priming" callback and it will
now see them as soon as it comes online.  A test is added for this case.

Change-Id: Iabb62cb060c2397ebca4e3ff1499b434e2da13df